+++ apps/chaingraph-frontend/src/App.tsx
@@ -6,7 +6,11 @@
  * As of the Change Date specified in that file, in accordance with the Business Source License, use of this software will be governed by the Apache License, version 2.0.
  */
 
+import { NodeRegistry } from '@badaitech/chaingraph-types'
 import { BrowserRouter, Navigate, Route, Routes } from 'react-router-dom'
+import SuperJSON from 'superjson'
+import { appConfig } from './config'
+import { RootProvider } from './exports'
 import { FlowLayout } from './FlowLayout'
 
 function App() {
@@ -14,8 +18,17 @@ function App() {
     <BrowserRouter>
       <Routes>
         <Route path="/" element={<Navigate to="/flows" replace />} />
-        <Route path="/flows" element={<FlowLayout />} />
-        <Route path="/flow/:flowId" element={<FlowLayout />} />
+        <Route element={(
+          <RootProvider
+            trpcURL={appConfig.chaingraphTRPCWSUrl}
+            superjsonCustom={SuperJSON}
+            nodeRegistry={NodeRegistry.getInstance()}
+          />
+        )}
+        >
+          <Route path="/flows" element={<FlowLayout />} />
+          <Route path="/flow/:flowId" element={<FlowLayout />} />
+        </Route>
       </Routes>
     </BrowserRouter>
   )