+++ apps/chaingraph-frontend/src/components/flow/nodes/ChaingraphNode/ports/utils/hide-editor.ts
@@ -6,7 +6,7 @@
  * As of the Change Date specified in that file, in accordance with the Business Source License, use of this software will be governed by the Apache License, version 2.0.
  */
 
-import type { EdgeData } from '@/store'
+import type { EdgeData } from '@/store/edges'
 import type { IPortConfig } from '@badaitech/chaingraph-types'
 
 export function isHideEditor(config: IPortConfig, connectedEdges: EdgeData[]) {