+++ apps/chaingraph-frontend/src/components/ui/context-menu.tsx
@@ -11,7 +11,7 @@ import * as ContextMenuPrimitive from '@radix-ui/react-context-menu'
 import { Check, ChevronRight, Circle } from 'lucide-react'
 
 import * as React from 'react'
-import { useShadowRoot } from './shadow'
+import { useShadowRoot } from './useShadowRoot'
 
 const ContextMenu = ContextMenuPrimitive.Root