+++ apps/chaingraph-frontend/src/components/ui/dialog.tsx
@@ -11,7 +11,7 @@ import * as DialogPrimitive from '@radix-ui/react-dialog'
 import { X } from 'lucide-react'
 
 import * as React from 'react'
-import { useShadowRoot } from './shadow'
+import { useShadowRoot } from './useShadowRoot'
 
 const Dialog = DialogPrimitive.Root