+++ apps/chaingraph-frontend/src/components/ui/select.tsx
@@ -11,7 +11,7 @@ import * as SelectPrimitive from '@radix-ui/react-select'
 import { Check, ChevronDown, ChevronUp } from 'lucide-react'
 
 import * as React from 'react'
-import { useShadowRoot } from './shadow'
+import { useShadowRoot } from './useShadowRoot'
 
 const Select = SelectPrimitive.Root